KARAF-7889: Upgrade to Apache POM 35 and target Java SE 11
