+++ itests/test/src/test/java/org/apache/karaf/itests/mavenresolver/MavenResolverRegisteredOnlyAfterConfigAdminTest.java
@@ -22,6 +22,7 @@ import static org.ops4j.pax.exam.karaf.options.KarafDistributionOption.editConfi
 
 import java.io.File;
 
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.ops4j.pax.exam.Configuration;
@@ -55,6 +56,7 @@ public class MavenResolverRegisteredOnlyAfterConfigAdminTest extends KarafMinima
     }
 
     @Test
+    @Ignore("Tinybundles should be updated in KarafMinimalMonitoredTestSupport")
     public void mavenResolverAvailable() throws Exception {
         long count = numberOfServiceEventsFor("org.ops4j.pax.url.mvn.MavenResolver");
         assertEquals("There should be only one MavenResolver registration - after non-INITIAL ConfigAdmin update", 1l, count);
