+++ specs/java.xml/pom.xml
@@ -31,6 +31,10 @@
     <packaging>jar</packaging>
     <name>Apache Karaf :: Specs :: Java Xml</name>
 
+    <properties>
+        <maven.compiler.target>8</maven.compiler.target>
+    </properties>
+
     <dependencyManagement>
         <dependencies>
             <dependency>
