+++ specs/locator/pom.xml
@@ -30,4 +30,8 @@
     <artifactId>org.apache.karaf.specs.locator</artifactId>
     <name>Apache Karaf :: Specs :: Locator</name>
 
+    <properties>
+        <maven.compiler.target>8</maven.compiler.target>
+    </properties>
+
 </project>
